fix assert caused by Rafiki's throttling
now that throttling causes delayed evaluation of messages there can be a reace condition where both, a message and the timeout are processed in the same iteration. Before we had guarrantee that timeouts are processed first but now the processing of a message is executed by the timer system such that we can't rely on this order anymore. I fixed this by ignoring timeouts for messages that were already processed.